完成学员端的 教练页面前后端联调和页面缺陷修改

+++ GymLink/src/main/java/com/ldr/gymlink/model/dto/course/CourseQueryPageRequest.java
@@ -25,7 +25,7 @@ public class CourseQueryPageRequest extends PageRequest implements Serializable 
     /**
      * 关联教练ID
      */
-    private Integer coachId;
+    private Long coachId;
 
     /**
      * 难度等级
